+++ opnieuw/retries.py
@@ -10,6 +10,8 @@ import logging
 import random
 import time
 from collections import defaultdict
+from contextlib import contextmanager
+from contextvars import ContextVar
 from typing import (
     Awaitable,
     cast,
@@ -137,9 +139,34 @@ class RetryState:
             )
 
 
-__retry_state_namespaces: Dict[Optional[str], Type[RetryState]] = defaultdict(
-    lambda: RetryState
-)
+__retry_state_namespaces: Dict[
+    Optional[str], ContextVar[Type[RetryState]]
+] = defaultdict(lambda: ContextVar("opnieuw_default_retry_state", default=RetryState))
+
+
+def _get_retry_state_class(namespace: Optional[str]) -> Type[RetryState]:
+    return __retry_state_namespaces[namespace].get()
+
+
+@contextmanager
+def replace_retry_state(
+    state: Type[RetryState], *, namespace: Optional[str] = None
+) -> Iterator[None]:
+    """
+    A context manager that replaces the state of the specified namespace with the
+    given `RetryState`.
+    This can be useful to customize the retry behavior, such as disabling the sleep interval during
+    tests (see `opnieuw.test_util.retry_immediately`).
+
+    Note: the retry state is context-local, meaning that changing the state in a thread or
+    asyncio task will not bleed to other threads or asyncio tasks.
+    See https://docs.python.org/3/library/contextvars.html for more details.
+    """
+    token = __retry_state_namespaces[namespace].set(state)
+    try:
+        yield
+    finally:
+        __retry_state_namespaces[namespace].reset(token)
 
 
 def retry(
@@ -202,13 +229,14 @@ def retry(
     Opnieuw is based on a retry algorithm off of:
         https://aws.amazon.com/blogs/architecture/exponential-backoff-and-jitter/
     """
+
     def decorator(f: F) -> F:
         @functools.wraps(f)
         def wrapper(*args: Any, **kwargs: Any) -> Any:
 
             last_exception = None
 
-            retry_state = __retry_state_namespaces[namespace](
+            retry_state = _get_retry_state_class(namespace)(
                 MonotonicClock(),
                 max_calls_total=max_calls_total,
                 retry_window_after_first_call_in_seconds=retry_window_after_first_call_in_seconds,
@@ -273,7 +301,7 @@ def retry_async(
 
             last_exception = None
 
-            retry_state = __retry_state_namespaces[namespace](
+            retry_state = _get_retry_state_class(namespace)(
                 MonotonicClock(),
                 max_calls_total=max_calls_total,
                 retry_window_after_first_call_in_seconds=retry_window_after_first_call_in_seconds,